Parser: Improved Parser, Added performance tests
