+++ src/main/kotlin/io/github/elaralang/elara/lexer/TokenType.kt
@@ -34,7 +34,7 @@ enum class TokenType(val regex: Regex) {
     COLON(":".toRegex()),
     SLASH("/".toRegex()),
 
-    IDENTIFIER("[^,.{}\\[\\]\"\\s)(]+".toRegex()); //this has the potential to get *very* messy...
+    IDENTIFIER("[^,.#{}\\[\\]\"\\s)(]+".toRegex()); //this has the potential to get *very* messy...
 
     companion object {
         // Creating regex to capture tokens
